divide file for place api logic
